requirements.txt
torch>=2.2.0
transformers==4.44.2
accelerate==0.34.2
trl==0.9.6
peft==0.12.0
datasets==2.20.0
evaluate==0.4.2
bitsandbytes==0.43.3
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
pyyaml>=6.0.1
scikit-learn>=1.3.0
# API Mistral et Backend
mistralai>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
aiofiles>=23.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pydantic-settings>=2.0.0
# Development tools
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
ruff>=0.1.0
black>=23.0.0
python-dotenv>=1.0.0
# Job queue
celery>=5.3.0
redis>=5.0.0
# Object storage
boto3>=1.34.0
